Remove Footprint Layer (and organize footprint functions) via #289
