+++ costmap_2d/cfg/ObstaclePlugin.cfg
@@ -5,6 +5,7 @@ from dynamic_reconfigure.parameter_generator_catkin import ParameterGenerator, b
 gen = ParameterGenerator()
 
 gen.add("enabled", bool_t, 0, "Whether to apply this plugin or not", True)
+gen.add("footprint_clearing_enabled", bool_t, 0, "Whether to clear the robot's footprint of lethal obstacles", True)
 gen.add("max_obstacle_height", double_t, 0, "The maximum height of any obstacle to be inserted into the costmap in meters.", 2, 0, 50)
 
 combo_enum = gen.enum([ gen.const("Overwrite", int_t,  0, "Overwrite values"),
