+++ costmap_2d/cfg/VoxelPlugin.cfg
@@ -5,6 +5,7 @@ from dynamic_reconfigure.parameter_generator_catkin import ParameterGenerator, b
 gen = ParameterGenerator()
 
 gen.add("enabled", bool_t, 0, "Whether to use this plugin or not", True)
+gen.add("footprint_clearing_enabled", bool_t, 0, "Whether to clear the robot's footprint of lethal obstacles", True)
 gen.add("max_obstacle_height", double_t, 0, "Max Obstacle Height", 2.0, 0, 50)
 gen.add("origin_z", double_t, 0, "The z origin of the map in meters.", 0, 0)
 gen.add("z_resolution", double_t, 0, "The z resolution of the map in meters/cell.", 0.2, 0, 50)
