+++ costmap_2d/src/footprint.cpp
@@ -32,6 +32,7 @@
 #include <boost/foreach.hpp>
 #include <boost/algorithm/string.hpp>
 #include <costmap_2d/footprint.h>
+#include <costmap_2d/array_parser.h>
 #include<geometry_msgs/Point32.h>
 
 namespace costmap_2d
@@ -101,5 +102,224 @@ std::vector<geometry_msgs::Point> toPointVector(geometry_msgs::Polygon polygon)
   }
   return pts;
 }
+
+void transformFootprint(double x, double y, double theta, const std::vector<geometry_msgs::Point>& footprint_spec, std::vector<geometry_msgs::Point>& oriented_footprint) {
+  //build the oriented footprint at a given location
+  oriented_footprint.clear();
+  double cos_th = cos(theta);
+  double sin_th = sin(theta);
+  for(unsigned int i = 0; i < footprint_spec.size(); ++i){
+    geometry_msgs::Point new_pt;
+    new_pt.x = x + (footprint_spec[i].x * cos_th - footprint_spec[i].y * sin_th);
+    new_pt.y = y + (footprint_spec[i].x * sin_th + footprint_spec[i].y * cos_th);
+    oriented_footprint.push_back(new_pt);
+  }
+}
+
+void transformFootprint(double x, double y, double theta, const std::vector<geometry_msgs::Point>& footprint_spec, geometry_msgs::PolygonStamped& oriented_footprint) {
+  //build the oriented footprint at a given location
+  oriented_footprint.polygon.points.clear();
+  double cos_th = cos(theta);
+  double sin_th = sin(theta);
+  for(unsigned int i = 0; i < footprint_spec.size(); ++i){
+    geometry_msgs::Point32 new_pt;
+    new_pt.x = x + (footprint_spec[i].x * cos_th - footprint_spec[i].y * sin_th);
+    new_pt.y = y + (footprint_spec[i].x * sin_th + footprint_spec[i].y * cos_th);
+    oriented_footprint.polygon.points.push_back(new_pt);
+  }
+}
+
+void padFootprint( std::vector<geometry_msgs::Point>& footprint, double padding )
+{
+  // pad footprint in place
+  for( unsigned int i = 0; i < footprint.size(); i++ )
+  {
+    geometry_msgs::Point& pt = footprint[ i ];
+    pt.x += sign( pt.x ) * padding;
+    pt.y += sign( pt.y ) * padding;
+  }
+}
+
+
+std::vector<geometry_msgs::Point> makeFootprintFromRadius( double radius )
+{
+  std::vector<geometry_msgs::Point> points;
+
+  // Loop over 16 angles around a circle making a point each time
+  int N = 16;
+  geometry_msgs::Point pt;
+  for( int i = 0; i < N; ++i )
+  {
+    double angle = i * 2 * M_PI / N;
+    pt.x = cos( angle ) * radius;
+    pt.y = sin( angle ) * radius;
+
+    points.push_back( pt );
+  }
+
+  return points;
+}
+
+
+std::vector<geometry_msgs::Point> makeFootprintFromString( const std::string& footprint_string, bool* error_flag )
+{
+  std::string error;
+  std::vector<std::vector<float> > vvf = parseVVF( footprint_string, error );
+  std::vector<geometry_msgs::Point> points;
+  
+  if( error != "" )
+  {
+    ROS_ERROR( "Error parsing footprint parameter: '%s'", error.c_str() );
+    ROS_ERROR( "  Footprint string was '%s'.", footprint_string.c_str() );
+    *error_flag = true;
+    return points;
+  }
+
+  // convert vvf into points.
+  if( vvf.size() < 3 )
+  {
+    ROS_ERROR( "You must specify at least three points for the robot footprint, reverting to previous footprint." );
+    *error_flag = true;
+    return points;
+  }
+  points.reserve( vvf.size() );
+  for( unsigned int i = 0; i < vvf.size(); i++ )
+  {
+    if( vvf[ i ].size() == 2 )
+    {
+      geometry_msgs::Point point;
+      point.x = vvf[ i ][ 0 ];
+      point.y = vvf[ i ][ 1 ];
+      point.z = 0;
+      points.push_back( point );
+    }
+    else
+    {
+      ROS_ERROR( "Points in the footprint specification must be pairs of numbers.  Found a point with %d numbers.",
+                 int( vvf[ i ].size() ));
+      *error_flag = true;
+      return points;
+    }
+  }
+
+  *error_flag = false;
+  return points;
+}
+
+
+
+std::vector<geometry_msgs::Point> makeFootprintFromParams( ros::NodeHandle& nh )
+{
+  std::string full_param_name;
+  std::string full_radius_param_name;
+  std::vector<geometry_msgs::Point> points; 
+
+  if( nh.searchParam( "footprint", full_param_name ))
+  {
+    XmlRpc::XmlRpcValue footprint_xmlrpc;
+    nh.getParam( full_param_name, footprint_xmlrpc );
+    if( footprint_xmlrpc.getType() == XmlRpc::XmlRpcValue::TypeString )
+    {
+      bool error_flag;
+      points = makeFootprintFromString( std::string( footprint_xmlrpc ), &error_flag );
+      if(error_flag)
+      {
+        writeFootprintToParam( nh, points);
+      }
+    }
+    else if( footprint_xmlrpc.getType() == XmlRpc::XmlRpcValue::TypeArray )
+    {
+      points = makeFootprintFromXMLRPC( footprint_xmlrpc, full_param_name );
+      writeFootprintToParam( nh, points );
+    }
+  }
+  else if( nh.searchParam( "robot_radius", full_radius_param_name ))
+  {
+    double robot_radius;
+    nh.param( full_radius_param_name, robot_radius, 1.234 );
+    points = makeFootprintFromRadius( robot_radius );
+    nh.setParam( "robot_radius", robot_radius );
+  }
+  // Else neither param was found anywhere this knows about, so
+  // defaults will come from dynamic_reconfigure stuff, set in
+  // cfg/Costmap2D.cfg and read in this file in reconfigureCB().
+  return points;
+}
+
+void writeFootprintToParam( ros::NodeHandle& nh, const std::vector<geometry_msgs::Point>& footprint)
+{
+  std::ostringstream oss;
+  bool first = true;
+  for( unsigned int i = 0; i < footprint.size(); i++ )
+  {
+    geometry_msgs::Point p = footprint[ i ];
+    if( first )
+    {
+      oss << "[[" << p.x << "," << p.y << "]";
+      first = false;
+    }
+    else
+    {
+      oss << ",[" << p.x << "," << p.y << "]";
+    }
+  }
+  oss << "]";
+  nh.setParam( "footprint", oss.str().c_str() );
+}
+
+double getNumberFromXMLRPC( XmlRpc::XmlRpcValue& value, const std::string& full_param_name )
+{
+  // Make sure that the value we're looking at is either a double or an int.
+  if( value.getType() != XmlRpc::XmlRpcValue::TypeInt &&
+      value.getType() != XmlRpc::XmlRpcValue::TypeDouble )
+  {
+    std::string& value_string = value;
+    ROS_FATAL( "Values in the footprint specification (param %s) must be numbers. Found value %s.",
+               full_param_name.c_str(), value_string.c_str() );
+    throw std::runtime_error("Values in the footprint specification must be numbers");
+  }
+  return value.getType() == XmlRpc::XmlRpcValue::TypeInt ? (int)(value) : (double)(value);
+}
+
+std::vector<geometry_msgs::Point> makeFootprintFromXMLRPC( XmlRpc::XmlRpcValue& footprint_xmlrpc,
+                                const std::string& full_param_name )
+{
+  // Make sure we have an array of at least 3 elements.
+  if( footprint_xmlrpc.getType() != XmlRpc::XmlRpcValue::TypeArray ||
+      footprint_xmlrpc.size() < 3 )
+  {
+    ROS_FATAL( "The footprint must be specified as list of lists on the parameter server, %s was specified as %s",
+               full_param_name.c_str(), std::string( footprint_xmlrpc ).c_str() );
+    throw std::runtime_error( "The footprint must be specified as list of lists on the parameter server with at least 3 points eg: [[x1, y1], [x2, y2], ..., [xn, yn]]");
+  }
+
+  std::vector<geometry_msgs::Point> footprint;  
+  geometry_msgs::Point pt;
+
+  for( int i = 0; i < footprint_xmlrpc.size(); ++i )
+  {
+    // Make sure each element of the list is an array of size 2. (x and y coordinates)
+    XmlRpc::XmlRpcValue point = footprint_xmlrpc[ i ];
+    if( point.getType() != XmlRpc::XmlRpcValue::TypeArray ||
+        point.size() != 2 )
+    {
+      ROS_FATAL( "The footprint (parameter %s) must be specified as list of lists on the parameter server eg: [[x1, y1], [x2, y2], ..., [xn, yn]], but this spec is not of that form.",
+                 full_param_name.c_str() );
+      throw std::runtime_error( "The footprint must be specified as list of lists on the parameter server eg: [[x1, y1], [x2, y2], ..., [xn, yn]], but this spec is not of that form" );
+    }
+
+    pt.x = getNumberFromXMLRPC( point[ 0 ], full_param_name );
+    pt.y = getNumberFromXMLRPC( point[ 1 ], full_param_name );
+
+    footprint.push_back( pt );
+  }
+  return footprint;
+}
+
+double sign(double x){
+  return x < 0.0 ? -1.0 : (x > 0.0 ? 1.0 : 0.0);
+}
+
+
 
 } // end namespace costmap_2d
